Add configuration keys to enable metrics
- Moved metrics classes to `dhis-support-system`
- Code cleanup
