+++ dhis-2/dhis-services/dhis-service-core/pom.xml
@@ -190,17 +190,6 @@
       <artifactId>hamcrest-library</artifactId>
       <scope>test</scope>
     </dependency>
-
-    <!-- application monitoring -->
-    <dependency>
-      <groupId>io.micrometer</groupId>
-      <artifactId>micrometer-spring-legacy</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.micrometer</groupId>
-      <artifactId>micrometer-registry-prometheus</artifactId>
-    </dependency>
-
   </dependencies>
   <properties>
     <rootDir>../../</rootDir>
