+++ dhis-2/dhis-support/dhis-support-external/src/main/java/org/hisp/dhis/external/conf/ConfigurationKey.java
@@ -91,7 +91,14 @@ public enum ConfigurationKey
     LOGGING_ADAPTER_CONSOLE_FORMAT( "logging.console.format", "TEXT" ),
     LOGGING_FILE_MAX_SIZE( "logging.file.max_size", "100MB" ),
     LOGGING_FILE_MAX_ARCHIVES( "logging.file.max_archives", "0" ),
-    SERVER_BASE_URL( "server.base.url", "", false );
+    SERVER_BASE_URL( "server.base.url", "", false ),
+    // MONITORING //
+    MONITORING_PROVIDER( "monitoring.provider", "prometheus" ),
+    MONITORING_API_ENABLED( "monitoring.api.enabled", "false", false ),
+    MONITORING_JVM_ENABLED( "monitoring.jvm.enabled", "false", false ),
+    MONITORING_DBPOOL_ENABLED( "monitoring.dbpool.enabled", "false", false ),
+    MONITORING_TOMCAT_ENABLED( "monitoring.tomcat.enabled", "false", false ),
+    MONITORING_HIBERNATE_ENABLED( "monitoring.hibernate.enabled", "false", false );
 
     private final String key;
