+++ dhis-2/dhis-support/dhis-support-system/src/main/java/org/hisp/dhis/monitoring/metrics/jdbc/DataSourcePoolMetadataProviders.java
@@ -26,31 +26,48 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.webapi.config;
+package org.hisp.dhis.monitoring.metrics.jdbc;
 
-import com.google.common.collect.Lists;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
 
-import io.micrometer.core.instrument.MeterRegistry;
-import io.micrometer.core.instrument.binder.tomcat.TomcatMetrics;
+import javax.sql.DataSource;
 
 /**
- * @author Luciano Fiandesio
+ * A {@link DataSourcePoolMetadataProvider} implementation that returns the first
+ * {@link DataSourcePoolMetadata} that is found by one of its delegate.
+ *
+ * @author Stephane Nicoll
+ * @since 1.2.0
  */
-@Configuration
-public class TomcatMetricsConfig
-{
-    @Bean
-    public TomcatMetrics tomcatMetrics()
-    {
-        return new TomcatMetrics( null, Lists.newArrayList() );
+public class DataSourcePoolMetadataProviders implements DataSourcePoolMetadataProvider {
+
+    private final List<DataSourcePoolMetadataProvider> providers;
+
+    /**
+     * Create a {@link DataSourcePoolMetadataProviders} instance with an initial
+     * collection of delegates to use.
+     * @param providers the data source pool metadata providers
+     */
+    public DataSourcePoolMetadataProviders(
+            Collection<? extends DataSourcePoolMetadataProvider> providers) {
+        this.providers = (providers == null
+                ? Collections.emptyList()
+                : new ArrayList<>(providers));
     }
 
-    @Autowired
-    public void bindToRegistry( MeterRegistry registry, TomcatMetrics tomcatMetrics )
-    {
-        tomcatMetrics.bindTo( registry );
+    @Override
+    public DataSourcePoolMetadata getDataSourcePoolMetadata(DataSource dataSource) {
+        for (DataSourcePoolMetadataProvider provider : this.providers) {
+            DataSourcePoolMetadata metadata = provider
+                    .getDataSourcePoolMetadata(dataSource);
+            if (metadata != null) {
+                return metadata;
+            }
+        }
+        return null;
     }
+
 }
